refactor(test): P5-4 extract loop + scan/projects from test_cli_shellout
Phase 5 P5-4 first cut. Three deliverables:

1. test_cli_shellout_helpers.hpp — shared header with run_pulp,
   pulp_binary, binary_exists, unique_temp_dir, ScopedEnvVar,
   write_text/read_file, prepend_to_path, pinned_shipyard_version_for_test,
   write_fake_shipyard. Previously inline in the parent's anonymous
   namespace.

2. test_cli_shellout_loop.cpp (10 TEST_CASEs, issue #940) — pulp loop
   --help / --status / --off / --platform / --no-watch /
   --watch-issues / --ar-swap-from / parser validation. Run with
   PULP_HOME pointed at a per-test tmpdir.

3. test_cli_shellout_scan_projects.cpp (12 TEST_CASEs) — pulp scan
   --help / --no-load / --format filesystem + HOME-derived CLAP +
   lv2 rich-scanner + single-bucket restriction + parser validation;
   pulp projects list --json / human-text / missing_on_disk / parser
   validation.

test_cli_shellout.cpp: 3,439 → 2,646 (-793 lines, closer to the
< 2,000 line P5-4 target).
