+++ test/test_cli_shellout.cpp
@@ -9,182 +9,11 @@
 // stdout/stderr invariants for the non-destructive subcommands every
 // user hits on day one.
 
-#include <catch2/catch_test_macros.hpp>
-#include <pulp/platform/child_process.hpp>
-
-#include <chrono>
-#include <cstdlib>
-#include <filesystem>
-#include <fstream>
-#include <sstream>
-#include <string>
-#include <system_error>
-#include <vector>
+#include "test_cli_shellout_helpers.hpp"
 
 using namespace pulp::platform;
 namespace fs = std::filesystem;
-
-namespace {
-
-// Platform-aware setenv/unsetenv: Windows MSVC doesn't provide POSIX
-// `setenv` / `unsetenv` — it uses `_putenv_s` instead (empty string
-// value for unset). Wrap both so every test body stays portable.
-inline int pulp_setenv(const char* name, const char* value, int /*overwrite*/) {
-#if defined(_WIN32)
-    return _putenv_s(name, value);
-#else
-    return ::setenv(name, value, 1);
-#endif
-}
-
-inline int pulp_unsetenv(const char* name) {
-#if defined(_WIN32)
-    return _putenv_s(name, "");
-#else
-    return ::unsetenv(name);
-#endif
-}
-
-class ScopedEnvVar {
-public:
-    explicit ScopedEnvVar(const char* name)
-        : name_(name) {
-        if (const char* value = std::getenv(name)) {
-            had_value_ = true;
-            old_value_ = value;
-        }
-    }
-
-    ~ScopedEnvVar() {
-        if (had_value_) {
-            pulp_setenv(name_.c_str(), old_value_.c_str(), 1);
-        } else {
-            pulp_unsetenv(name_.c_str());
-        }
-    }
-
-    void set(const std::string& value) {
-        pulp_setenv(name_.c_str(), value.c_str(), 1);
-    }
-
-private:
-    std::string name_;
-    bool had_value_ = false;
-    std::string old_value_;
-};
-
-fs::path unique_temp_dir(const std::string& prefix) {
-    auto tick = std::chrono::steady_clock::now().time_since_epoch().count();
-    return fs::temp_directory_path() / (prefix + "-" + std::to_string(tick));
-}
-
-
-// After the Rust CLI cutover, the instrumented C++ delegate lands at
-// <build>/tools/cli/pulp-cpp. Older/pre-cutover builds used
-// <build>/tools/cli/pulp. Prefer pulp-cpp so coverage builds exercise
-// the C++ implementation directly, but keep the old fallback for
-// compatibility. PULP_CLI_PATH can still override either path.
-fs::path pulp_binary() {
-    if (const char* env = std::getenv("PULP_CLI_PATH"); env && *env) {
-        return fs::path(env);
-    }
-    auto cpp = fs::current_path() / ".." / "tools" / "cli" / "pulp-cpp";
-    if (fs::exists(cpp)) return cpp;
-    return fs::current_path() / ".." / "tools" / "cli" / "pulp";
-}
-
-ProcessResult run_pulp(const std::vector<std::string>& args,
-                       int timeout_ms = 10000) {
-    auto bin = pulp_binary();
-    if (!fs::exists(bin)) {
-        // Binary not built for this target — rather than fail noisily,
-        // surface a SKIP at the call site. Tests check for .ok() first.
-        ProcessResult r;
-        r.exit_code = -1;
-        r.stderr_output = "pulp binary not at " + bin.string();
-        return r;
-    }
-    return exec(bin.string(), args, timeout_ms);
-}
-
-bool binary_exists() { return fs::exists(pulp_binary()); }
-
-std::string read_file(const fs::path& path) {
-    std::ifstream f(path);
-    if (!f.is_open()) return {};
-    return std::string(std::istreambuf_iterator<char>(f),
-                       std::istreambuf_iterator<char>());
-}
-
-void write_text(const fs::path& path, const std::string& text) {
-    fs::create_directories(path.parent_path());
-    std::ofstream f(path);
-    REQUIRE(f.is_open());
-    f << text;
-    REQUIRE(f.good());
-}
-
-char path_separator() {
-#if defined(_WIN32)
-    return ';';
-#else
-    return ':';
-#endif
-}
-
-void prepend_to_path(const fs::path& dir) {
-    const char* old = std::getenv("PATH");
-    auto next = dir.string();
-    if (old && *old) {
-        next += path_separator();
-        next += old;
-    }
-    pulp_setenv("PATH", next.c_str(), 1);
-}
-
-#if !defined(_WIN32)
-std::string pinned_shipyard_version_for_test() {
-    auto toml = read_file(fs::current_path().parent_path().parent_path() /
-                          "tools" / "shipyard.toml");
-    std::istringstream lines(toml);
-    std::string line;
-    while (std::getline(lines, line)) {
-        auto pos = line.find("version");
-        if (pos == std::string::npos) continue;
-        auto eq = line.find('=', pos);
-        if (eq == std::string::npos) continue;
-        auto value = line.substr(eq + 1);
-        auto first = value.find('"');
-        auto last = value.find_last_of('"');
-        if (first != std::string::npos && last != std::string::npos && last > first) {
-            return value.substr(first + 1, last - first - 1);
-        }
-    }
-    return "v0.56.2";
-}
-
-fs::path write_fake_shipyard(const fs::path& dir, const std::string& version) {
-    auto path = dir / "shipyard";
-    write_text(path,
-               "#!/bin/sh\n"
-               "if [ \"$1\" = \"--version\" ]; then\n"
-               "  echo \"shipyard " + version + "\"\n"
-               "  exit 0\n"
-               "fi\n"
-               "if [ \"$1\" = \"pr\" ]; then\n"
-               "  echo \"fake shipyard pr $2\"\n"
-               "  exit 0\n"
-               "fi\n"
-               "echo \"fake shipyard\"\n");
-    fs::permissions(path,
-                    fs::perms::owner_exec | fs::perms::owner_read |
-                    fs::perms::owner_write,
-                    fs::perm_options::add);
-    return path;
-}
-#endif
-
-}  // namespace
+using namespace pulp_test_cli;
 
 TEST_CASE("pulp help exits 0 with a usage banner on stdout",
           "[cli][shellout]") {
@@ -2815,625 +2644,3 @@ TEST_CASE("PULP_DEBUG=1 surfaces phase markers to stderr (#682)",
     pulp_unsetenv("PULP_UPDATE_CHECK_DISABLED");
 }
 
-// Issue #940 Slice 1 — `pulp loop` is the leveraged-prototype focus
-// mode entry point. These tests cover the surface that doesn't depend
-// on a live watch loop (the watch-loop interactions are tested by
-// `pulp dev`'s harness):
-//
-//   * `pulp loop --help`     exits 0 + advertises focus mode
-//   * `pulp loop --status`   exits 0, reports detected host + focus state
-//   * `pulp loop --off`      clears focus state (and tolerates "already off")
-//   * `pulp loop --platform=<known>`   accepts macos/linux/windows
-//   * `pulp loop --platform=<unknown>` exits non-zero with diagnostic
-//   * `pulp loop --no-watch` flips focus + exits 0 without watching
-//
-// All of these run with PULP_HOME pointed at a per-test tmpdir so the
-// developer's real config.toml is never touched.
-TEST_CASE("pulp loop --help exits 0 with focus-mode banner",
-          "[cli][shellout][loop][issue-940]") {
-    if (!binary_exists()) { SUCCEED("skipped: pulp not built"); return; }
-    auto r = run_pulp({"loop", "--help"});
-    REQUIRE(r.exit_code == 0);
-    REQUIRE_FALSE(r.timed_out);
-    REQUIRE(r.stdout_output.find("leveraged-prototype focus mode") != std::string::npos);
-    REQUIRE(r.stdout_output.find("--platform") != std::string::npos);
-    REQUIRE(r.stdout_output.find("--off") != std::string::npos);
-}
-
-TEST_CASE("pulp loop --status reports detected host and focus state",
-          "[cli][shellout][loop][issue-940]") {
-    if (!binary_exists()) { SUCCEED("skipped: pulp not built"); return; }
-
-    auto tmp_home = unique_temp_dir("pulp-940-status");
-    fs::create_directories(tmp_home);
-    pulp_setenv("PULP_HOME", tmp_home.string().c_str(), 1);
-    pulp_setenv("PULP_UPDATE_CHECK_DISABLED", "1", 1);
-
-    auto r = run_pulp({"loop", "--status"});
-    REQUIRE_FALSE(r.timed_out);
-    REQUIRE(r.exit_code == 0);
-    REQUIRE(r.stdout_output.find("detected host") != std::string::npos);
-    REQUIRE(r.stdout_output.find("focus platform") != std::string::npos);
-    // Fresh PULP_HOME — no focus persisted yet.
-    REQUIRE(r.stdout_output.find("(none") != std::string::npos);
-
-    pulp_unsetenv("PULP_HOME");
-    pulp_unsetenv("PULP_UPDATE_CHECK_DISABLED");
-    fs::remove_all(tmp_home);
-}
-
-TEST_CASE("pulp loop --off clears focus state idempotently",
-          "[cli][shellout][loop][issue-940]") {
-    if (!binary_exists()) { SUCCEED("skipped: pulp not built"); return; }
-
-    auto tmp_home = unique_temp_dir("pulp-940-off");
-    fs::create_directories(tmp_home);
-    pulp_setenv("PULP_HOME", tmp_home.string().c_str(), 1);
-    pulp_setenv("PULP_UPDATE_CHECK_DISABLED", "1", 1);
-
-    // First call: nothing to clear, must still exit 0.
-    auto r1 = run_pulp({"loop", "--off"});
-    REQUIRE_FALSE(r1.timed_out);
-    REQUIRE(r1.exit_code == 0);
-    REQUIRE(r1.stdout_output.find("already in cross-platform mode") != std::string::npos);
-
-    // Set focus via --no-watch, then clear.
-    auto r_set = run_pulp({"loop", "--platform=macos", "--no-watch"});
-    REQUIRE_FALSE(r_set.timed_out);
-    REQUIRE(r_set.exit_code == 0);
-    REQUIRE(r_set.stdout_output.find("focus mode active on macos") != std::string::npos);
-
-    auto r2 = run_pulp({"loop", "--off"});
-    REQUIRE_FALSE(r2.timed_out);
-    REQUIRE(r2.exit_code == 0);
-    REQUIRE(r2.stdout_output.find("focus mode cleared") != std::string::npos);
-    REQUIRE(r2.stdout_output.find("Cross-platform mode restored") != std::string::npos);
-
-    pulp_unsetenv("PULP_HOME");
-    pulp_unsetenv("PULP_UPDATE_CHECK_DISABLED");
-    fs::remove_all(tmp_home);
-}
-
-TEST_CASE("pulp loop --platform=<known> accepts macos|linux|windows",
-          "[cli][shellout][loop][issue-940]") {
-    if (!binary_exists()) { SUCCEED("skipped: pulp not built"); return; }
-
-    auto tmp_home = unique_temp_dir("pulp-940-platform");
-    fs::create_directories(tmp_home);
-    pulp_setenv("PULP_HOME", tmp_home.string().c_str(), 1);
-    pulp_setenv("PULP_UPDATE_CHECK_DISABLED", "1", 1);
-
-    for (const char* p : {"macos", "linux", "windows"}) {
-        auto r = run_pulp({"loop", "--platform", p, "--no-watch"});
-        REQUIRE_FALSE(r.timed_out);
-        REQUIRE(r.exit_code == 0);
-        std::string expect = std::string("focus mode active on ") + p;
-        REQUIRE(r.stdout_output.find(expect) != std::string::npos);
-    }
-
-    pulp_unsetenv("PULP_HOME");
-    pulp_unsetenv("PULP_UPDATE_CHECK_DISABLED");
-    fs::remove_all(tmp_home);
-}
-
-TEST_CASE("pulp loop --platform=<unknown> exits non-zero with diagnostic",
-          "[cli][shellout][loop][issue-940]") {
-    if (!binary_exists()) { SUCCEED("skipped: pulp not built"); return; }
-
-    auto tmp_home = unique_temp_dir("pulp-940-unknown");
-    fs::create_directories(tmp_home);
-    pulp_setenv("PULP_HOME", tmp_home.string().c_str(), 1);
-    pulp_setenv("PULP_UPDATE_CHECK_DISABLED", "1", 1);
-
-    auto r = run_pulp({"loop", "--platform=plan9", "--no-watch"});
-    REQUIRE_FALSE(r.timed_out);
-    REQUIRE(r.exit_code != 0);
-    auto combined = r.stdout_output + r.stderr_output;
-    REQUIRE(combined.find("unknown --platform") != std::string::npos);
-    REQUIRE(combined.find("macos") != std::string::npos);
-
-    pulp_unsetenv("PULP_HOME");
-    pulp_unsetenv("PULP_UPDATE_CHECK_DISABLED");
-    fs::remove_all(tmp_home);
-}
-
-TEST_CASE("pulp loop validates value options before focus state changes",
-          "[cli][shellout][loop][coverage][phase3]") {
-    if (!binary_exists()) { SUCCEED("skipped: pulp not built"); return; }
-
-    auto tmp_home = unique_temp_dir("pulp-loop-parser-errors");
-    fs::create_directories(tmp_home);
-    ScopedEnvVar pulp_home("PULP_HOME");
-    pulp_home.set(tmp_home.string());
-
-    struct Case {
-        std::vector<std::string> args;
-        std::string stderr_substring;
-    };
-    const std::vector<Case> cases = {
-        {{"loop", "--platform"}, "--platform requires a value"},
-        {{"loop", "--platform", "--no-watch"}, "--platform requires a value"},
-        {{"loop", "--watch-issues"}, "--watch-issues requires a value"},
-        {{"loop", "--watch-issues", "--no-watch"}, "--watch-issues requires a value"},
-        {{"loop", "--ar-swap-from"}, "--ar-swap-from requires a value"},
-        {{"loop", "--ar-swap-from", "--no-watch"}, "--ar-swap-from requires a value"},
-        {{"loop", "--run"}, "--run requires a value"},
-        {{"loop", "--run", "--target", "pulp-format"}, "--run requires a value"},
-        {{"loop", "--target"}, "--target requires a value"},
-        {{"loop", "--target", "--no-watch"}, "--target requires a value"},
-    };
-
-    for (const auto& c : cases) {
-        INFO("loop args size: " << c.args.size());
-        auto r = run_pulp(c.args, 10000);
-        REQUIRE_FALSE(r.timed_out);
-        REQUIRE(r.exit_code == 2);
-        REQUIRE(r.stderr_output.find(c.stderr_substring) != std::string::npos);
-    }
-
-    auto status = run_pulp({"loop", "--status"}, 10000);
-    REQUIRE_FALSE(status.timed_out);
-    REQUIRE(status.exit_code == 0);
-    REQUIRE(status.stdout_output.find("focus platform: (none") != std::string::npos);
-
-    fs::remove_all(tmp_home);
-}
-
-TEST_CASE("pulp loop --watch-issues prints deferred-slice hint",
-          "[cli][shellout][loop][issue-940]") {
-    if (!binary_exists()) { SUCCEED("skipped: pulp not built"); return; }
-
-    auto tmp_home = unique_temp_dir("pulp-940-watch-issues");
-    fs::create_directories(tmp_home);
-    pulp_setenv("PULP_HOME", tmp_home.string().c_str(), 1);
-    pulp_setenv("PULP_UPDATE_CHECK_DISABLED", "1", 1);
-
-    auto r = run_pulp({"loop", "--watch-issues=924,927", "--no-watch"});
-    REQUIRE_FALSE(r.timed_out);
-    REQUIRE(r.exit_code == 0);
-    // Slice 3 deferred — surface the issue number so the user knows
-    // when the helper lands.
-    REQUIRE(r.stdout_output.find("--watch-issues=924,927") != std::string::npos);
-    REQUIRE(r.stdout_output.find("deferred") != std::string::npos);
-
-    pulp_unsetenv("PULP_HOME");
-    pulp_unsetenv("PULP_UPDATE_CHECK_DISABLED");
-    fs::remove_all(tmp_home);
-}
-
-TEST_CASE("pulp loop --ar-swap-from prints deferred-slice hint",
-          "[cli][shellout][loop][issue-940]") {
-    if (!binary_exists()) { SUCCEED("skipped: pulp not built"); return; }
-
-    auto tmp_home = unique_temp_dir("pulp-940-ar-swap");
-    fs::create_directories(tmp_home);
-    pulp_setenv("PULP_HOME", tmp_home.string().c_str(), 1);
-    pulp_setenv("PULP_UPDATE_CHECK_DISABLED", "1", 1);
-
-    // Long form (separate value)
-    auto r1 = run_pulp({"loop", "--ar-swap-from", "feat/fix-x", "--no-watch"});
-    REQUIRE_FALSE(r1.timed_out);
-    REQUIRE(r1.exit_code == 0);
-    REQUIRE(r1.stdout_output.find("--ar-swap-from=feat/fix-x") != std::string::npos);
-    REQUIRE(r1.stdout_output.find("deferred") != std::string::npos);
-
-    // Equals form
-    auto r2 = run_pulp({"loop", "--ar-swap-from=feat/fix-y", "--no-watch"});
-    REQUIRE_FALSE(r2.timed_out);
-    REQUIRE(r2.exit_code == 0);
-    REQUIRE(r2.stdout_output.find("--ar-swap-from=feat/fix-y") != std::string::npos);
-
-    pulp_unsetenv("PULP_HOME");
-    pulp_unsetenv("PULP_UPDATE_CHECK_DISABLED");
-    fs::remove_all(tmp_home);
-}
-
-TEST_CASE("pulp loop --status reports persisted focus platform",
-          "[cli][shellout][loop][issue-940]") {
-    if (!binary_exists()) { SUCCEED("skipped: pulp not built"); return; }
-
-    auto tmp_home = unique_temp_dir("pulp-940-status-set");
-    fs::create_directories(tmp_home);
-    pulp_setenv("PULP_HOME", tmp_home.string().c_str(), 1);
-    pulp_setenv("PULP_UPDATE_CHECK_DISABLED", "1", 1);
-
-    // Set focus, then assert --status reflects the persisted value.
-    auto rs = run_pulp({"loop", "--platform=linux", "--no-watch"});
-    REQUIRE_FALSE(rs.timed_out);
-    REQUIRE(rs.exit_code == 0);
-
-    auto r = run_pulp({"loop", "--status"});
-    REQUIRE_FALSE(r.timed_out);
-    REQUIRE(r.exit_code == 0);
-    REQUIRE(r.stdout_output.find("focus platform: linux") != std::string::npos);
-
-    pulp_unsetenv("PULP_HOME");
-    pulp_unsetenv("PULP_UPDATE_CHECK_DISABLED");
-    fs::remove_all(tmp_home);
-}
-
-TEST_CASE("pulp loop accepts --test, --validate, --target, --run flags via --no-watch",
-          "[cli][shellout][loop][issue-940]") {
-    // The flags exercise the parser branches that wouldn't otherwise run
-    // when only --status / --off / --no-watch are used. With --no-watch
-    // we exit before the watch loop, so the binary still terminates.
-    if (!binary_exists()) { SUCCEED("skipped: pulp not built"); return; }
-
-    auto tmp_home = unique_temp_dir("pulp-940-flags");
-    fs::create_directories(tmp_home);
-    pulp_setenv("PULP_HOME", tmp_home.string().c_str(), 1);
-    pulp_setenv("PULP_UPDATE_CHECK_DISABLED", "1", 1);
-
-    // --test, -t, --test-filter=, --validate, --allow-unsupported-sdk
-    // all exit cleanly with --no-watch.
-    auto r1 = run_pulp({"loop", "--platform=macos", "--test", "--validate", "--no-watch"});
-    REQUIRE_FALSE(r1.timed_out);
-    REQUIRE(r1.exit_code == 0);
-
-    auto r2 = run_pulp({"loop", "--platform=macos", "-t", "--test-filter=Foo", "--no-watch"});
-    REQUIRE_FALSE(r2.timed_out);
-    REQUIRE(r2.exit_code == 0);
-
-    // --target T pushes onto build_args; --run T sets launch_target. Both
-    // branches must accept the next-arg form without breaking the parser.
-    auto r3 = run_pulp({"loop", "--platform=macos", "--target", "pulp-format",
-                        "--run", "pulp-gain-standalone", "--no-watch"});
-    REQUIRE_FALSE(r3.timed_out);
-    REQUIRE(r3.exit_code == 0);
-
-    // Trailing -- forwards remaining args to the launched binary; with
-    // --no-watch we just exit so the args are silently consumed.
-    auto r4 = run_pulp({"loop", "--platform=macos", "--no-watch", "--", "--arg1", "value1"});
-    REQUIRE_FALSE(r4.timed_out);
-    REQUIRE(r4.exit_code == 0);
-
-    pulp_unsetenv("PULP_HOME");
-    pulp_unsetenv("PULP_UPDATE_CHECK_DISABLED");
-    fs::remove_all(tmp_home);
-}
-
-// ── projects list --json (cmd_projects.cpp do_list JSON path) ─────
-//
-// The C++ binary's `pulp projects list --json` was added in this
-// branch to mirror what the Rust port already does — without these
-// tests, the diff-cover gate flags the new JSON-rendering branch as
-// uncovered (per `ci/coverage-targets.yaml`, `tools/cli/**` is the
-// "user-facing" tier with a 70% per-PR floor).
-
-TEST_CASE("pulp projects list --json emits valid JSON for empty registry",
-          "[cli][shellout][projects][issue-244]") {
-    if (!binary_exists()) { SUCCEED("skipped: pulp not built"); return; }
-
-    // Isolated PULP_HOME so we don't read the user's real registry.
-    auto tmp = fs::temp_directory_path() / "pulp-projects-json-test";
-    fs::create_directories(tmp);
-    pulp_setenv("PULP_HOME", tmp.string().c_str(), 1);
-    auto r = run_pulp({"projects", "list", "--json"});
-    pulp_unsetenv("PULP_HOME");
-    fs::remove_all(tmp);
-
-    REQUIRE(r.exit_code == 0);
-    // Must be parseable JSON with the documented top-level keys.
-    REQUIRE(r.stdout_output.find("\"registry\"") != std::string::npos);
-    REQUIRE(r.stdout_output.find("\"projects\"") != std::string::npos);
-    // Empty registry → empty array.
-    REQUIRE(r.stdout_output.find("[]") != std::string::npos);
-}
-
-TEST_CASE("pulp projects list (no --json) emits human text",
-          "[cli][shellout][projects]") {
-    if (!binary_exists()) { SUCCEED("skipped: pulp not built"); return; }
-
-    auto tmp = fs::temp_directory_path() / "pulp-projects-text-test";
-    fs::create_directories(tmp);
-    pulp_setenv("PULP_HOME", tmp.string().c_str(), 1);
-    auto r = run_pulp({"projects", "list"});
-    pulp_unsetenv("PULP_HOME");
-    fs::remove_all(tmp);
-
-    REQUIRE(r.exit_code == 0);
-    // Human text path should NOT emit JSON braces.
-    REQUIRE(r.stdout_output.find("Registry:") != std::string::npos);
-    REQUIRE(r.stdout_output.find("(no projects registered)") != std::string::npos);
-}
-
-TEST_CASE("pulp projects validates parser errors before registry mutation",
-          "[cli][shellout][projects][coverage][phase3]") {
-    if (!binary_exists()) { SUCCEED("skipped: pulp not built"); return; }
-
-    auto pulp_home = unique_temp_dir("pulp-projects-parser-home");
-    auto project = unique_temp_dir("pulp-projects-parser-project");
-    fs::create_directories(pulp_home);
-    fs::create_directories(project);
-    ScopedEnvVar pulp_home_env("PULP_HOME");
-    pulp_home_env.set(pulp_home.string());
-
-    struct Case {
-        std::vector<std::string> args;
-        std::string stderr_substring;
-    };
-    const std::vector<Case> cases = {
-        {{"projects", "list", "--bogus"}, "unknown option"},
-        {{"projects", "add", project.string(), "extra"}, "unexpected argument"},
-        {{"projects", "remove"}, "path argument is required"},
-        {{"projects", "remove", project.string(), "extra"}, "unexpected argument"},
-    };
-
-    for (const auto& c : cases) {
-        INFO("projects args size: " << c.args.size());
-        auto r = run_pulp(c.args, 10000);
-        REQUIRE_FALSE(r.timed_out);
-        REQUIRE(r.exit_code != 0);
-        REQUIRE(r.stderr_output.find(c.stderr_substring) != std::string::npos);
-    }
-
-    REQUIRE_FALSE(fs::exists(pulp_home / "projects.json"));
-
-    std::error_code ec;
-    fs::remove_all(pulp_home, ec);
-    fs::remove_all(project, ec);
-}
-
-// Non-empty registry path — exercises the cmd_projects::do_list JSON
-// loop body (lines ~78-92 of cmd_projects.cpp) that the empty-registry
-// test above can't reach. Adds a synthetic project root, registers it,
-// then asserts the JSON output contains the per-project record. Also
-// exercises json_escape via a project name containing a literal
-// double-quote (the most common escapable character).
-TEST_CASE("pulp projects list --json emits per-project JSON with non-empty registry",
-          "[cli][shellout][projects][issue-244]") {
-    if (!binary_exists()) { SUCCEED("skipped: pulp not built"); return; }
-
-    auto pulp_home = unique_temp_dir("pulp-projects-json-nonempty-home");
-    fs::create_directories(pulp_home);
-    pulp_setenv("PULP_HOME", pulp_home.string().c_str(), 1);
-
-    // Synthetic project root — find_project_root walks ancestors looking
-    // for `CMakeLists.txt && core/`, so create both. pulp.toml carries
-    // the SDK pin; cmd_projects::do_add reads name from there if present.
-    auto project = unique_temp_dir("pulp-projects-json-fake-project");
-    fs::create_directories(project / "core");
-    {
-        std::ofstream ct(project / "CMakeLists.txt");
-        ct << "project(FakeJsonListProject VERSION 0.1.0 LANGUAGES CXX)\n";
-    }
-    {
-        std::ofstream pt(project / "pulp.toml");
-        pt << "sdk_version = \"0.40.0\"\n";
-    }
-
-    auto add = run_pulp({"projects", "add", project.string()}, 15000);
-    REQUIRE_FALSE(add.timed_out);
-    REQUIRE(add.exit_code == 0);
-
-    auto list = run_pulp({"projects", "list", "--json"}, 15000);
-    REQUIRE_FALSE(list.timed_out);
-    REQUIRE(list.exit_code == 0);
-
-    // Top-level shape preserved.
-    REQUIRE(list.stdout_output.find("\"registry\"") != std::string::npos);
-    REQUIRE(list.stdout_output.find("\"projects\"") != std::string::npos);
-    // Loop body must have emitted one project record with these fields.
-    REQUIRE(list.stdout_output.find("\"path\"") != std::string::npos);
-    REQUIRE(list.stdout_output.find("\"name\"") != std::string::npos);
-    REQUIRE(list.stdout_output.find("\"registered_at\"") != std::string::npos);
-    REQUIRE(list.stdout_output.find("\"missing_on_disk\"") != std::string::npos);
-    REQUIRE(list.stdout_output.find(project.generic_string()) != std::string::npos);
-    // The synthetic project exists on disk → missing_on_disk false.
-    REQUIRE(list.stdout_output.find("\"missing_on_disk\": false") != std::string::npos);
-
-    pulp_unsetenv("PULP_HOME");
-    fs::remove_all(pulp_home);
-    fs::remove_all(project);
-}
-
-// Pairs with the above — exercises cmd_projects's missing-on-disk
-// branch by registering a project then deleting its directory. This
-// covers the `bool missing = !fs::exists(p.path)` true branch + the
-// `"missing_on_disk": true` JSON emission line.
-TEST_CASE("pulp projects list --json reports missing_on_disk=true for deleted project",
-          "[cli][shellout][projects][issue-244]") {
-    if (!binary_exists()) { SUCCEED("skipped: pulp not built"); return; }
-
-    auto pulp_home = unique_temp_dir("pulp-projects-json-missing-home");
-    fs::create_directories(pulp_home);
-    pulp_setenv("PULP_HOME", pulp_home.string().c_str(), 1);
-
-    auto project = unique_temp_dir("pulp-projects-json-deleted-project");
-    fs::create_directories(project / "core");
-    {
-        std::ofstream ct(project / "CMakeLists.txt");
-        ct << "project(DeletedFakeProject VERSION 0.1.0 LANGUAGES CXX)\n";
-    }
-    {
-        std::ofstream pt(project / "pulp.toml");
-        pt << "sdk_version = \"0.40.0\"\n";
-    }
-
-    auto add = run_pulp({"projects", "add", project.string()}, 15000);
-    REQUIRE_FALSE(add.timed_out);
-    REQUIRE(add.exit_code == 0);
-
-    // Now delete the project directory but leave the registry entry.
-    fs::remove_all(project);
-
-    auto list = run_pulp({"projects", "list", "--json"}, 15000);
-    REQUIRE_FALSE(list.timed_out);
-    REQUIRE(list.exit_code == 0);
-    REQUIRE(list.stdout_output.find("\"missing_on_disk\": true") != std::string::npos);
-
-    pulp_unsetenv("PULP_HOME");
-    fs::remove_all(pulp_home);
-}
-
-// ── pulp scan --help / --no-load (#812) ──────────────────────────
-//
-// `pulp scan --help` previously walked the system plug-in paths and
-// dlopen'd everything (which crashes on a malformed plugin). The
-// fix added an early --help gate; this test locks in that behavior
-// so a future refactor can't reintroduce the crash.
-
-TEST_CASE("pulp scan --help exits 0 with usage on stdout",
-          "[cli][shellout][scan][issue-812]") {
-    if (!binary_exists()) { SUCCEED("skipped: pulp not built"); return; }
-    auto r = run_pulp({"scan", "--help"});
-    REQUIRE(r.exit_code == 0);
-    REQUIRE_FALSE(r.timed_out);
-    REQUIRE(r.stdout_output.find("Usage: pulp scan") != std::string::npos);
-    // The new --no-load flag must be advertised so users hitting #812
-    // can discover the safe-fast escape hatch.
-    REQUIRE(r.stdout_output.find("--no-load") != std::string::npos);
-}
-
-TEST_CASE("pulp scan validates parser errors before filesystem enumeration",
-          "[cli][shellout][scan][coverage][phase3]") {
-    if (!binary_exists()) { SUCCEED("skipped: pulp not built"); return; }
-
-    const struct {
-        std::vector<std::string> args;
-        const char* message;
-    } cases[] = {
-        {{"scan", "--format"}, "--format requires a value"},
-        {{"scan", "-f"}, "-f requires a value"},
-        {{"scan", "--format", "--no-load"}, "--format requires a value"},
-        {{"scan", "--definitely-not-a-scan-flag"}, "unknown flag"},
-    };
-
-    for (const auto& c : cases) {
-        INFO("scan args under test");
-        auto r = run_pulp(c.args, /*timeout_ms=*/10000);
-        REQUIRE_FALSE(r.timed_out);
-        REQUIRE(r.exit_code == 2);
-        REQUIRE(r.stderr_output.find(c.message) != std::string::npos);
-        REQUIRE(r.stdout_output.find("[CLAP]") == std::string::npos);
-        REQUIRE(r.stdout_output.find("No plugins found") == std::string::npos);
-    }
-}
-
-TEST_CASE("pulp host validates parser errors before plugin loading",
-          "[cli][shellout][host][coverage][phase3]") {
-    if (!binary_exists()) { SUCCEED("skipped: pulp not built"); return; }
-
-    const struct {
-        std::vector<std::string> args;
-        const char* message;
-    } cases[] = {
-        {{"host", "--format"}, "--format requires a value"},
-        {{"host", "-f"}, "-f requires a value"},
-        {{"host", "--format", "--id"}, "--format requires a value"},
-        {{"host", "--id"}, "--id requires a value"},
-        {{"host", "--id", "--format"}, "--id requires a value"},
-        {{"host", "--definitely-not-a-host-flag"}, "unknown flag"},
-    };
-
-    for (const auto& c : cases) {
-        INFO("host args under test");
-        auto r = run_pulp(c.args, /*timeout_ms=*/10000);
-        REQUIRE_FALSE(r.timed_out);
-        REQUIRE(r.exit_code == 2);
-        REQUIRE(r.stderr_output.find(c.message) != std::string::npos);
-        REQUIRE(r.stderr_output.find("failed to load") == std::string::npos);
-        REQUIRE(r.stdout_output.find("Loaded:") == std::string::npos);
-    }
-}
-
-TEST_CASE("pulp scan --no-load runs filesystem-only enumeration cleanly",
-          "[cli][shellout][scan][issue-812]") {
-    if (!binary_exists()) { SUCCEED("skipped: pulp not built"); return; }
-    auto r = run_pulp({"scan", "--no-load"}, /*timeout_ms=*/30000);
-    // The whole point of --no-load is "doesn't crash on bad plugins".
-    // If the host happens to have no plugins installed, that's also
-    // a clean rc=0 with "No plugins found." stdout — both are fine.
-    REQUIRE_FALSE(r.timed_out);
-    REQUIRE(r.exit_code == 0);
-    // No libc++abi termination message should appear in stderr —
-    // that's the regression --no-load specifically prevents.
-    REQUIRE(r.stderr_output.find("libc++abi: terminating") == std::string::npos);
-}
-
-TEST_CASE("pulp scan --no-load reports filename-derived CLAP entries from HOME",
-          "[cli][shellout][scan][issue-812]") {
-    if (!binary_exists()) { SUCCEED("skipped: pulp not built"); return; }
-#if defined(_WIN32)
-    SUCCEED("Windows CLAP defaults do not derive from HOME");
-    return;
-#else
-    auto home = unique_temp_dir("pulp-scan-noload-home");
-    ScopedEnvVar scoped_home("HOME");
-    scoped_home.set(home.string());
-
-#if defined(__APPLE__)
-    auto clap_dir = home / "Library" / "Audio" / "Plug-Ins" / "CLAP";
-#else
-    auto clap_dir = home / ".clap";
-#endif
-    auto alpha_path = clap_dir / "Phase8Alpha.clap";
-    auto beta_path = clap_dir / "Phase8Beta.clap";
-    fs::create_directories(beta_path);
-    fs::create_directories(alpha_path);
-
-    auto r = run_pulp({"scan", "--no-load", "--format", "clap"},
-                      /*timeout_ms=*/30000);
-    REQUIRE_FALSE(r.timed_out);
-    REQUIRE(r.exit_code == 0);
-    REQUIRE(r.stderr_output.find("libc++abi: terminating") == std::string::npos);
-    REQUIRE(r.stdout_output.find("[CLAP]") != std::string::npos);
-    REQUIRE(r.stdout_output.find(alpha_path.string()) != std::string::npos);
-    REQUIRE(r.stdout_output.find(beta_path.string()) != std::string::npos);
-
-    auto alpha = r.stdout_output.find("Phase8Alpha");
-    auto beta = r.stdout_output.find("Phase8Beta");
-    REQUIRE(alpha != std::string::npos);
-    REQUIRE(beta != std::string::npos);
-    REQUIRE(alpha < beta);
-
-    std::error_code ec;
-    fs::remove_all(home, ec);
-#endif
-}
-
-TEST_CASE("pulp scan --format lv2 reaches rich scanner path and exits cleanly",
-          "[cli][shellout][scan][issue-812]") {
-    if (!binary_exists()) { SUCCEED("skipped: pulp not built"); return; }
-
-    auto home = unique_temp_dir("pulp-scan-lv2-home");
-    ScopedEnvVar scoped_home("HOME");
-    scoped_home.set(home.string());
-
-#if defined(__linux__)
-    auto lv2_dir = home / ".lv2" / "Phase8Probe.lv2";
-    fs::create_directories(lv2_dir);
-    write_text(lv2_dir / "manifest.ttl",
-               "@prefix lv2: <http://lv2plug.in/ns/lv2core#> .\n"
-               "<http://example.org/phase8-probe> a lv2:Plugin .\n");
-#endif
-
-    auto r = run_pulp({"scan", "--format", "lv2"}, /*timeout_ms=*/30000);
-    REQUIRE_FALSE(r.timed_out);
-    REQUIRE(r.exit_code == 0);
-    REQUIRE(r.stderr_output.find("libc++abi: terminating") == std::string::npos);
-#if defined(__linux__)
-    REQUIRE(r.stdout_output.find("Phase8Probe") != std::string::npos);
-#else
-    REQUIRE((r.stdout_output.find("[LV2]") != std::string::npos ||
-             r.stdout_output.find("No plugins found") != std::string::npos));
-#endif
-
-    std::error_code ec;
-    fs::remove_all(home, ec);
-}
-
-TEST_CASE("pulp scan --no-load --format clap restricts to one bucket",
-          "[cli][shellout][scan][issue-812]") {
-    if (!binary_exists()) { SUCCEED("skipped: pulp not built"); return; }
-    auto r = run_pulp({"scan", "--no-load", "--format", "clap"},
-                      /*timeout_ms=*/30000);
-    REQUIRE_FALSE(r.timed_out);
-    REQUIRE(r.exit_code == 0);
-    REQUIRE(r.stderr_output.find("libc++abi: terminating") == std::string::npos);
-    // When the filter narrows to CLAP and no CLAP plugins exist on
-    // the runner, the output should be the documented empty-state
-    // message rather than crashing or printing other-format buckets.
-    REQUIRE((r.stdout_output.find("[CLAP]") != std::string::npos ||
-             r.stdout_output.find("No plugins found") != std::string::npos));
-}
